# client/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : ""
  },
  "basePath" : "/",
  "schemes" : [ "http" ],
  "paths" : {
...
    "/individuals/{nedId}/addresses" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "List addresses",
        "description" : "",
        "operationId" : "getAddresses",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Address"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "individuals" ],
        "summary" : "Create address",
        "description" : "",
        "operationId" : "createAddress",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Address"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Address"
            }
          }
        }
      }
    },
    "/individuals/{nedId}/addresses/{addressId}" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "Read address",
        "description" : "",
        "operationId" : "getAddress",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "addressId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Address"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "individuals" ],
        "summary" : "Update address",
        "description" : "",
        "operationId" : "updateAddress",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "addressId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Address"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Address"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "individuals" ],
        "summary" : "Delete address",
        "description" : "",
        "operationId" : "deleteAddress",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "addressId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ]
      }
    },
...
    "/individuals/{nedId}/emails" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "List emails",
        "description" : "",
        "operationId" : "getEmails",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Email"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "individuals" ],
        "summary" : "Create email",
        "description" : "",
        "operationId" : "createEmail",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Email"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Email"
            }
          }
        }
      }
    },
    "/individuals/{nedId}/emails/{emailId}" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "Read email",
        "description" : "",
        "operationId" : "getEmail",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "emailId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Email"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "individuals" ],
        "summary" : "Update email",
        "description" : "",
        "operationId" : "updateEmail",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "emailId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Email"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Email"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "individuals" ],
        "summary" : "Delete email",
        "description" : "",
        "operationId" : "deleteEmail",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "emailId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ]
      }
    },
...
    "/individuals/{nedId}/phonenumbers" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "List phone numbers",
        "description" : "",
        "operationId" : "getPhonenumbers",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Phonenumber"
              }
            }
          }
        }
      }
    },
...
    "/individuals/{nedId}/uids" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "List UIDs",
        "description" : "",
        "operationId" : "getUids",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Uniqueidentifier"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "individuals" ],
        "summary" : "Create UID",
        "description" : "",
        "operationId" : "createUid",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Uniqueidentifier"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Uniqueidentifier"
            }
          }
        }
      }
    },
    "/individuals/{nedId}/uids/{id}" : {
      "get" : {
        "tags" : [ "individuals" ],
        "summary" : "Read uid",
        "description" : "",
        "operationId" : "getUid",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Uniqueidentifier"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "individuals" ],
        "summary" : "Update UID",
        "description" : "",
        "operationId" : "updateUid",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Uniqueidentifier"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Uniqueidentifier"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "individuals" ],
        "summary" : "Delete UID",
        "description" : "",
        "operationId" : "deleteUid",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ]
      }
    },
...
    "/organizations/{nedId}/addresses" : {
      "get" : {
        "tags" : [ "organizations" ],
        "summary" : "List addresses",
        "description" : "",
        "operationId" : "getAddresses",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Address"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "organizations" ],
        "summary" : "Create address",
        "description" : "",
        "operationId" : "createAddress",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Address"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Address"
            }
          }
        }
      }
    },
...
    "/organizations/{nedId}/emails" : {
      "get" : {
        "tags" : [ "organizations" ],
        "summary" : "List emails",
        "description" : "",
        "operationId" : "getEmails",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Email"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "organizations" ],
        "summary" : "Create email",
        "description" : "",
        "operationId" : "createEmail",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Email"
          }
        }, {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "$ref" : "#/definitions/Email"
            }
          }
        }
      }
    },
...
    "/organizations/{nedId}/phonenumbers" : {
      "get" : {
        "tags" : [ "organizations" ],
        "summary" : "List phone numbers",
        "description" : "",
        "operationId" : "getPhonenumbers",
        "parameters" : [ {
          "name" : "nedId",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Phonenumber"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "Address" : {
      "properties" : {
        "typeid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "type" : {
          "type" : "string"
        },
        "addressline1" : {
          "type" : "string"
        },
        "addressline2" : {
          "type" : "string"
        },
        "addressline3" : {
          "type" : "string"
        },
        "city" : {
          "type" : "string"
        },
        "statecodetypeid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "statecodetype" : {
          "type" : "string"
        },
        "countrycodetypeid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "countrycodetype" : {
          "type" : "string"
        },
        "postalcode" : {
          "type" : "string"
        },
        "maincontactnamedentityid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "latitude" : {
          "type" : "integer",
          "format" : "int32"
        },
        "longitude" : {
          "type" : "integer",
          "format" : "int32"
        },
        "isactive" : {
          "type" : "boolean"
        },
        "source" : {
          "type" : "string"
        },
        "nedid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sourcetypeid" : {
          "type" : "integer",
          "format" : "int32"
        },
        "id" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
  }
}